#75 Mock typed read
