+++ src/module/VisualizationParser.ts
@@ -17,6 +17,7 @@ import {
   IVisualization,
   IItem,
 } from '../../index';
+import { Mock } from './MockedTypeRead';
 
 export class VisualizationParser {
   public static parseEdgeList(
@@ -43,6 +44,7 @@ export class VisualizationParser {
     data: any;
     readAt: string;
   }): IOgmaEdge {
+    edge = Mock.properties(edge);
     return {
       id: edge.id,
       source: edge.source,
@@ -67,6 +69,7 @@ export class VisualizationParser {
     selected?: boolean;
     readAt: string;
   }): IOgmaNode {
+    node = Mock.properties(node);
     return {
       id: node.id,
       x: node.nodelink ? node.nodelink.x : undefined,
